+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Updated number of random videos showing on home page from 3 to 7
 - Updated 2 hero action button styles on home page
 - Added styles for news pagination link on home page
+- Added one more random video to a home page sidebar
 
 ----